bot/requirements.txt
python-telegram-bot>=20.0
anthropic>=0.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
